pyproject.toml
[tool.poetry]
name = "useful-habbit"
version = "0.1.0"
description = ""
authors = ["Антон Федяев <fearthe54@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.1"
pillow = "^10.3.0"
coverage = "^7.5.3"
requests = "^2.32.3"
django-celery-beat = "^2.6.0"
drf-yasg = "^1.21.7"
eventlet = "^0.36.1"
redis = "^5.0.5"
django-cors-headers = "^4.3.1"
python-dotenv = "^1.0.1"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
